+++ jms-db-domain/src/main/java/com/jms/domain/db/Groups.java
@@ -1,5 +1,5 @@
 package com.jms.domain.db;
-// Generated 2015-6-16 15:01:20 by Hibernate Tools 3.2.2.GA
+// Generated 2015-6-25 9:56:24 by Hibernate Tools 3.2.2.GA
 
 
 import java.util.Date;
@@ -31,7 +31,7 @@ import javax.persistence.TemporalType;
 public class Groups  implements java.io.Serializable {
 
 
-     private Integer idGroup;
+     private Long idGroup;
      private Users users;
      private String groupName;
      private String description;
@@ -58,11 +58,11 @@ public class Groups  implements java.io.Serializable {
      @Id @GeneratedValue(strategy=IDENTITY)
     
     @Column(name="ID_GROUP", unique=true, nullable=false)
-    public Integer getIdGroup() {
+    public Long getIdGroup() {
         return this.idGroup;
     }
     
-    public void setIdGroup(Integer idGroup) {
+    public void setIdGroup(Long idGroup) {
         this.idGroup = idGroup;
     }
 @ManyToOne(fetch=FetchType.LAZY)
