+++ jms-repositories/src/main/java/com/jms/repositories/company/SectorsRepository.java
@@ -13,13 +13,13 @@ import com.jms.domain.db.Sector;
 
 
 @Repository
-public interface SectorsRepository  extends CrudRepository<Sector, Integer>{
+public interface SectorsRepository  extends CrudRepository<Sector, Long>{
 	
 	@Query("select s from Sector s where s.sector=?1 and s.company.companyName=?2")	
 	public Sector findBySectorAndCompanyName(String sector,String companyName);
 
 	public List<Sector> findByCompany(Company company);
 	@Query("select s from Sector s where  s.company.idCompany=?1 order by seq")
-	public List<Sector> findByIdCompany(Integer idCompany);
+	public List<Sector> findByIdCompany(Long idCompany);
 
 }
